Fix: 로그아웃시 확인하도록 변경. 로그아웃 및 로그인 문구 변경.

+++ components/common/CommerceHeaderMenuModal.js
@@ -17,10 +17,12 @@ export default function CommerceHeaderMenuModal({
   const btnCss = { width: '100%' };
 
   const signout = () => {
-    localStorage.clear();
-    setToken('');
-    alert('로그아웃 완료');
-    router.push(LINKS.MAIN);
+    if (confirm('로그아웃 하시겠습니까?')) {
+      localStorage.clear();
+      setToken('');
+      alert('로그아웃 되었습니다.');
+      router.push(LINKS.MAIN);
+    }
   };
 
   function MainMenu() {
